requirements.txt
google-cloud-storage==2.13.0
streamlit==1.28.0
streamlit-authenticator==0.2.3
qrcode==7.4.2
pandas==2.1.2
